Scene with navigationBar, ImGui Layout configs
- camera can move, rotate, zoom, and rotate around fixed point in the
  distance
- Cinema4D-like imgui style
- more customizable window configurations which are saved and loaded from ini file
- shortcut in menuBar to get to every registered imgui window
- added native script based on TheCherno's Hazel
- new TextureAtlasUV (with many useful flags)
- fixed terrible memory leak during creation of atlases, (there was no
  free() after malloc()!
- possibility to load meshes as well as materials

- and many other things I honestly don't remember... btw what day is it?

+++ Sandbox/src/gui/MainWindow.cpp
@@ -527,8 +527,8 @@ ControlsWindow::ControlsWindow(const MessageConsumer& c)
 				if(eve.getEventType()==Event::EventType::KeyPress)
 				{
 					auto m = static_cast<KeyPressEvent&>(eve);
-					ControlMap::setValueAtPointer(titlo,m.getKey());
-					button->getTextElement()->setText(titlo + ": " + Font::colorize(Font::BLACK, "#013220") + ControlMap::getKeyName(m.getKey()));
+					ControlMap::setValueAtPointer(titlo,(uint64_t)m.getKey());
+					button->getTextElement()->setText(titlo + ": " + Font::colorize(Font::BLACK, "#013220") + ControlMap::getKeyName((uint64_t)m.getKey()));
 					GUIContext::get().setFocusedElement(nullptr);
 				}
 			}else if(eve.getEventType()==Event::EventType::MouseFocusGain)
@@ -573,8 +573,8 @@ ControlsWindow::ControlsWindow(const MessageConsumer& c)
 				if (eve.getEventType() == Event::EventType::KeyPress)
 				{
 					auto m = static_cast<KeyPressEvent&>(eve);
-					ControlMap::setValueAtPointer(titlo, m.getKey());
-					button->getTextElement()->setText(titlo + ": " + Font::colorize(Font::BLACK, "#013220") + ControlMap::getKeyName(m.getKey()));
+					ControlMap::setValueAtPointer(titlo, (uint64_t)m.getKey());
+					button->getTextElement()->setText(titlo + ": " + Font::colorize(Font::BLACK, "#013220") + ControlMap::getKeyName((uint64_t)m.getKey()));
 					GUIContext::get().setFocusedElement(nullptr);
 				}
 			}
